Use a more serializable type for oid block size.
`size_t' is _portable_ but not a great type for a wire protocol over a
network with heterogeneous platforms on it.

* gzochi-server/src/oids.h (gzochid_data_oids_block): `size_t' ->
  `guint64' for field `block_size'.

+++ gzochi-server/src/oids.h
@@ -50,7 +50,7 @@ enum GzochidOidError
 struct _gzochid_data_oids_block
 {
   guint64 block_start; /* The first object id in the block. */
-  size_t block_size; /* The size of the block. */
+  guint16 block_size; /* The size of the block. */
 };
 
 typedef struct _gzochid_data_oids_block gzochid_data_oids_block;
